+++ libraries/FastTrig/CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.3.2] - 2012-12-20
+- fix changelog
+- fix #25 change reference parameters to pointer 
+  - isincos(float f, float \*si, float \*co)
+  - isincos256(uint32_t v, int \*si, int \*co);
+  - fix examples
+
+
+## [0.3.1] - 2012-12-18
+- add hypotFast
+- add examples hypotFast
+- update readme.md
+
 ## [0.3.0] - 2012-12-11
 - add isin256(), icos256() and isincos256() => integer math only version
 - add isincos(float f, float &si, float &co) => calculate sin and cos simultaneously.